+++ tests/testthat/test-complete-lazy.R
@@ -102,4 +102,57 @@ test_that("argument 'explicit' works", {
   )
 })
 
+test_that("can use named arguments", {
+  df <- pl$LazyFrame(
+    group = c(1:2, 1, 2),
+    item_id = c(1:2, 2, 3),
+    item_name = c("a", "a", "b", "b"),
+    value1 = c(1, NA, 3, 4),
+    value2 = 4:7
+  )
+
+  expect_equal_lazy(
+    df |>
+      complete(group, value1 = c(1, 2, 3, 4)) |>
+      arrange(group, value1),
+    as.data.frame(df) |>
+      complete(group, value1 = c(1, 2, 3, 4)) |>
+      arrange(group, value1) |>
+      as.data.frame()
+  )
+
+  # only one named input
+  expect_equal_lazy(
+    df |>
+      complete(value1 = c(1, 2, 3, 4)) |>
+      arrange(value1),
+    as.data.frame(df) |>
+      complete(value1 = c(1, 2, 3, 4)) |>
+      arrange(value1) |>
+      as.data.frame()
+  )
+
+  # input columns are reordered
+  expect_equal_lazy(
+    df |>
+      complete(value1 = c(1, 2, 3, 4), group) |>
+      arrange(value1, group),
+    as.data.frame(df) |>
+      complete(value1 = c(1, 2, 3, 4), group) |>
+      arrange(value1, group) |>
+      as.data.frame()
+  )
+  expect_equal_lazy(
+    df |>
+      group_by(item_id) |>
+      complete(value1 = c(1, 2, 3, 4), group) |>
+      arrange(item_id, value1, group),
+    as.data.frame(df) |>
+      group_by(item_id) |>
+      complete(value1 = c(1, 2, 3, 4), group) |>
+      arrange(item_id, value1, group) |>
+      as.data.frame()
+  )
+})
+
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)